Fix #3080, test failures.
The failures were caused by a couple issues that may have been limited
to OSX (possibly only SL).
* The absence of /bin/true caused a couple Resolution.exec tests to
fail. Instead of expecting a binary to exist, the exists? check is now
stubbed.
* Changes to system_profiler output or the Resolution.exec result type
caused in an exception with the plist parser (which would cause
Factor.list, among other things, to fail); this is now handled gracefully.